// client/src/pages/Dashboard/types.ts
export interface Contact {
    id: string;
    name: string;
    cpf: string,
    email: string;
    phone: string;
    cep: number;
    logradouro: string;
    numero: string;
    complemento: string;
    bairro: string;
    cidade: string;
    estado: string;
    latitude: number,
    longitude: number,
  }